requirements.txt
# Google Cloud Storage dependencies
google-cloud-storage>=2.10.0

# Existing dependencies (add as needed)
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.0
tqdm>=4.62.0
transformers>=4.21.0
lmdb>=1.3.0
